Removed base_views almost entirely
