+++ backend/backend/urls.py
@@ -35,7 +35,6 @@ urlpatterns = [
     path(API_URL + "accounts/", include("accounts.urls")),
     path(API_URL + "blog-posts/", include("blogs.urls")),
     path(API_URL + "search/", include("search.urls")),
-    path(API_URL + "crossposts/", include("crossposts.urls")),
     path(
         "sitemap.xml",
         sitemap,
