+++ backend/blogs/views.py
@@ -1,4 +1,7 @@
+from django.db.models import F, Case, Count, ExpressionWrapper, FloatField, When
+from django.http.response import Http404
 from rest_framework import generics, status
+from rest_framework.parsers import FormParser, JSONParser, MultiPartParser
 from rest_framework.permissions import AllowAny, IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -6,19 +9,16 @@ from rest_framework.views import APIView
 from accounts.helpers import IsAdmin, IsModerator
 from accounts.models import CustomUser
 from accounts.serializers import CustomUserSerializer
-from base.base_views import (
-    BaseCommentView,
-    BaseFeedListView,
-    BasePostListView,
-    BasePostView,
-    BaseReactionView,
-)
-from blogs.models import BlogPost, PostComment, Follower
+from blogs.models import BlogPost, PostComment, Follower, PostReaction
 from blogs.serializers import (
     BlogPostSerializer,
+    CommentAndUserSerializer,
     CommentSerializer,
+    CreateOrUpdateBlogPostSerializer,
+    CreateOrUpdateCommentSerializer,
     FollowerSerializer,
     GetFollowerSerializer,
+    ReactionSerializer,
 )
 
 
@@ -35,13 +35,135 @@ class CommentListUserView(APIView):
 
 
 # this is the view for a *single* blog post
-class BlogPostView(BasePostView):
-    pass
+class BlogPostView(APIView):
+    parser_classes = (MultiPartParser, FormParser, JSONParser)
+
+    def get_permissions(self):
+        permissions = super().get_permissions()
+        if self.request.method.lower() != "get":  # type: ignore
+            permissions.append(IsAuthenticated())  # type: ignore
+        return permissions
+
+    def get(self, request, post_type, username, slug):
+        try:
+            res = (
+                BlogPost.objects.filter(post_type=post_type)
+                .select_related("user")
+                .annotate(likes=Count("postreaction"))
+                .get(user__username=username, slug_field=slug)
+            )
+            serializer = BlogPostSerializer(res)
+            return Response(data=serializer.data, status=status.HTTP_200_OK)
+        except BlogPost.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+    def post(self, request, post_type):
+        data = request.data
+
+        # blog post specific data
+        user = self.request.user.id  # type: ignore
+        title = data["title"]
+        content = data["content"]
+        thumbnail = data["thumbnail"]
+
+        if thumbnail == "undefined":
+            thumbnail = None
+        serializer_data = {
+            "user": user,
+            "title": title,
+            "content": content,
+            "thumbnail": thumbnail,
+            "post_type": post_type,
+        }
+
+        serializer = CreateOrUpdateBlogPostSerializer(data=serializer_data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
+
+        return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def put(self, request, post_type):
+        data = request.data
+
+        # post specific data
+        user = self.request.user.id  # type: ignore
+        title = data["title"]
+        title_slug = data["title_slug"]
+        content = data["content"]
+        thumbnail = data["thumbnail"]
+        original_slug = data["original_slug"]
+
+        try:
+            is_original_post = BlogPost.objects.get(
+                user=self.request.user, slug_field=title_slug
+            )
+        except BlogPost.DoesNotExist:
+            is_original_post = None
+
+        # weird conditional, just means that if the title has changed and theres a post with the title "title_slug", then return a 400
+        if title_slug != original_slug and is_original_post:
+            return Response(
+                data={"error": "A post with this title already exists!"},
+                status=status.HTTP_400_BAD_REQUEST,
+            )
+
+        if thumbnail == "undefined" and is_original_post:
+            thumbnail = is_original_post.thumbnail
+            if not thumbnail:
+                thumbnail = None
+        elif thumbnail == "undefined":
+            thumbnail = None
+
+        serializer_data = {
+            "user": user,
+            "title": title,
+            "content": content,
+            "thumbnail": thumbnail,
+            "post_type": post_type,
+        }
+
+        instance = generics.get_object_or_404(
+            BlogPost, user=user, slug_field=original_slug
+        )
+        serializer = CreateOrUpdateBlogPostSerializer(
+            data=serializer_data, instance=instance
+        )
+        if serializer.is_valid():
+            serializer.save()
+            return Response(data=serializer.data, status=status.HTTP_200_OK)
+
+        return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request):
+        data = request.data
+        user = self.request.user
+        to_delete = generics.get_object_or_404(
+            BlogPost, user=user, slug_field=data["slug"]
+        )
+        to_delete.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 # this is the view for *multiple* blog posts
-class BlogPostListView(BasePostListView):
-    pass
+class BlogPostListView(APIView):
+    permission_classes = (AllowAny,)
+
+    def get(self, request, username=None):
+        if username:
+            try:
+                res = BlogPost.objects.filter(user__username=username)
+                if len(res) == 0:
+                    raise BlogPost.DoesNotExist
+                serializer = BlogPostSerializer(res, many=True)
+            except BlogPost.DoesNotExist:
+                raise Http404
+            return Response(data=serializer.data, status=status.HTTP_200_OK)
+        else:
+            user = self.request.user.id  # type:ignore
+            res = BlogPost.objects.filter(user=user).select_related("user")
+            serializer = BlogPostSerializer(res, many=True)
+            return Response(data=serializer.data, status=status.HTTP_200_OK)
 
 
 # This view returns replies to *comments* that a user has made
@@ -54,8 +176,72 @@ class CommentReplyListView(generics.ListAPIView):
         return PostComment.objects.filter(user=user).exclude(reply_to=None)
 
 
-class BlogPostCommentView(BaseCommentView):
-    pass
+class BlogPostCommentView(APIView):
+    def get_permissions(self):
+        permissions = super().get_permissions()
+        if self.request.method.lower() != "get":  # type: ignore
+            permissions.append(IsAuthenticated())  # type: ignore
+        return permissions
+
+    def get(self, request, username, slug):
+        post = generics.get_object_or_404(
+            BlogPost, slug_field=slug, user__username=username
+        )
+        queryset = PostComment.objects.filter(post=post).select_related(
+            "user", "reply_to"
+        )
+        serializer = CommentAndUserSerializer(queryset, many=True)
+        return Response(data=serializer.data, status=status.HTTP_200_OK)
+
+    def post(self, request):
+        data = request.data
+
+        # this is mainly for the frontend, because it is being a pain.
+        if "reply_to" not in data.keys():
+            reply_to = None
+        else:
+            reply_to = data["reply_to"]
+
+        if reply_to == "":
+            reply_to = None
+
+        new_comment = {
+            "user": self.request.user.id,  # type:ignore
+            "post": generics.get_object_or_404(
+                BlogPost,
+                slug_field=data["slug"],
+                user__username=data["username"],
+            ).pk,
+            "content": data["content"],
+            "is_read": False,
+            "reply_to": reply_to,
+        }
+        serializer = CreateOrUpdateCommentSerializer(data=new_comment)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_201_CREATED)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    # in this case, updating a comment only involves changing the content
+    def patch(self, request, id):
+        data = request.data
+        user = self.request.user
+        comment = generics.get_object_or_404(PostComment, pk=id, user=user)
+        serializer = CreateOrUpdateCommentSerializer(
+            instance=comment, data={"content": data["content"], "user": user.id}  # type: ignore
+        )
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data, status=status.HTTP_200_OK)
+
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request, id):
+        user = self.request.user
+        comment = generics.get_object_or_404(PostComment, user=user, pk=id)
+        comment.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 # This view returns replies to *posts* that a user has made
@@ -68,8 +254,76 @@ class PostReplyListView(generics.ListAPIView):
         return PostComment.objects.filter(user=user, reply_to=None)
 
 
-class FeedListView(BaseFeedListView):
-    pass
+class FeedListView(APIView):
+    posts_per_page = 50
+    follower_addition = 50
+    listicle_debuff = 30
+    comment_score = 5
+    reaction_score = 3
+
+    def get(self, request, index):
+        index = int(index)
+        posts_per_page = self.posts_per_page
+
+        # id like to change this to a multiplier at some point
+        follower_addition = self.follower_addition
+        listicle_debuff = self.listicle_debuff
+        comment_score = self.comment_score
+        reaction_score = self.reaction_score
+
+        # initial query
+        all_posts = (
+            BlogPost.objects.all()
+            .annotate(
+                reactions=Count("postreaction"),
+                comments=Count("postcomment"),
+                score=ExpressionWrapper(
+                    (F("reactions") * reaction_score) + (F("comments") * comment_score),
+                    output_field=FloatField(),
+                ),
+            )
+            .select_related("user")
+            .order_by("-score")
+        )
+
+        # paginating list
+        all_posts = all_posts[posts_per_page * (index - 1) : posts_per_page * index]
+
+        # adding listicle "debuff"
+        all_posts = all_posts.annotate(
+            score=Case(
+                When(
+                    is_listicle=True,
+                    then=ExpressionWrapper(
+                        F("score") - listicle_debuff, output_field=FloatField()
+                    ),
+                ),
+                default=ExpressionWrapper(F("score"), output_field=FloatField()),
+            )
+        )
+
+        # adding follower "boosts"
+        if request.user.id:
+            following = (
+                Follower.objects.filter(follower=request.user.id)
+                .select_related("user")
+                .values("user__username")
+            )
+
+            all_posts = all_posts.annotate(
+                score=Case(
+                    When(
+                        user__username__in=following,
+                        then=ExpressionWrapper(
+                            F("score") + follower_addition, output_field=FloatField()
+                        ),
+                    ),
+                    default=ExpressionWrapper(F("score"), output_field=FloatField()),
+                ),
+            )
+
+        serializer = BlogPostSerializer(all_posts, many=True)
+        return Response(data=serializer.data, status=status.HTTP_200_OK)
 
 
 class FollowerView(APIView):
@@ -134,8 +388,42 @@ class FollowingView(APIView):
 
 
 # defaults for basereactionview are correct for PostReactionView
-class PostReactionView(BaseReactionView):
-    pass
+class PostReactionView(APIView):
+    permission_classes = (IsAuthenticated,)
+
+    # methods takes in user (self.request.user) and the post slug (its unique)
+    def get(self, request, slug, username):
+        # here, user refers to the authenticated user making the request, and username refers to the username of the user that published the article
+        user = self.request.user
+        blog_post = generics.get_object_or_404(
+            BlogPost, slug_field=slug, user__username=username
+        )
+        reaction = generics.get_object_or_404(PostReaction, post=blog_post, user=user)
+        serializer = ReactionSerializer(reaction)
+        return Response(data=serializer.data, status=status.HTTP_200_OK)
+
+    def post(self, request):
+        user = self.request.user.id  # type: ignore
+        slug = request.data["slug"]
+        username = request.data["username"]
+        blog_post = generics.get_object_or_404(
+            BlogPost, slug_field=slug, user__username=username
+        ).pk
+        data = {"user": user, "post": blog_post}
+        serializer = ReactionSerializer(data=data)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(data=serializer.data, status=status.HTTP_201_CREATED)
+
+        return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+    def delete(self, request):
+        user = self.request.user
+        slug = request.data["slug"]
+        blog_post = generics.get_object_or_404(BlogPost, slug_field=slug)
+        reaction = generics.get_object_or_404(PostReaction, post=blog_post, user=user)
+        reaction.delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 class ModeratorModifyPostView(APIView):
